Fix default config for WS
The default attrib value sets the user and pass. If using OAuth tokens
for WS, the default user and pass are still written to the config
file. Since the plugin checks first on the presence of user and pass
the OAuth tokens are never used and no valid auth can be done against
API. This commit waits to set the default config to the setup recipe
so the default config is only written if no config is provided.